Add files via upload
The project runs only on the console. When you run the code it asks you to choose one of the two options: admin/user. If you are the admin you can add new words to the database. And if you choose the user option you can search words from the database.